pyproject.toml
# =============================
# project build system settings
# =============================
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]  # required tools for buidling the package
build-backend = "setuptools.build_meta"  # backend used for buidling the package

# ======================
# project metadata
# ======================
[project]
name = "nomad-battery-database"  # name of the project
description = "app for battery database"  # short discription of project
version = "0.1.0"  # project verison
readme = "README.md"  # path to readme file
requires-python = ">=3.9"  # minimum required python version
license = { file = "LICENSE" }  # license file

# classifiers help categorize the project on pypi
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

# author and maintainer information
authors = [
    { name = "Uday Gajera", email = "uday.gajera@physik.hu-berlin.de" },
]
maintainers = [
    { name = "Uday Gajera", email = "uday.gajera@physik.hu-berlin.de" },
]
# ======================
# project dependencies
# ======================
dependencies = [
    "nomad-lab>=1.3.0",  # nomad lab integrataion
    "python-magic-bin; sys_platform == 'win32'",  # platform-specific dependancy (windows only)
    "pytest"  # testing framework
]

[project.urls]
Repository = "https://github.com/u-gajera/Battery-Database"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

# ======================
# nomad plugin entry points
# ======================
[project.entry-points.'nomad.plugin']
parser_entry_point = "nomad_battery_database.parsers.parser:BatteryParser"
schema_package_entry_point = "nomad_battery_database.schema_packages.schema_package:BatteryProperties"
#app_entry_point = "nomad_battery_database.apps:app_entry_point"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
